learnx-backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
SQLAlchemy
passlib[bcrypt]
python-jose[cryptography]
python-multipart
httpx
orjson
